Cancel local timestamp caching logic
